gun/rifle: move mesh manipulation to libtrx
