+++ src/libtrx/include/libtrx/game/gun/rifle.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "../lara/types.h"
+
+void Gun_Rifle_DrawMeshes(LARA_GUN_TYPE weapon_type);
+void Gun_Rifle_UndrawMeshes(LARA_GUN_TYPE weapon_type);
