+++ src/tr1/game/gun/gun_rifle.c
@@ -83,20 +83,6 @@ void Gun_Rifle_Undraw(const LARA_GUN_TYPE weapon_type)
     g_Lara.left_arm.frame_num = ani;
 }
 
-void Gun_Rifle_DrawMeshes(const LARA_GUN_TYPE weapon_type)
-{
-    Gun_SetLaraHandLMesh(weapon_type);
-    Gun_SetLaraHandRMesh(weapon_type);
-    Gun_SetLaraBackMesh(LGT_UNARMED);
-}
-
-void Gun_Rifle_UndrawMeshes(const LARA_GUN_TYPE weapon_type)
-{
-    Gun_SetLaraHandLMesh(LGT_UNARMED);
-    Gun_SetLaraHandRMesh(LGT_UNARMED);
-    Gun_SetLaraBackMesh(weapon_type);
-}
-
 void Gun_Rifle_Ready(const LARA_GUN_TYPE weapon_type)
 {
     g_Lara.gun_status = LGS_READY;
