+++ src/tr1/game/gun/gun_rifle.h
@@ -7,10 +7,10 @@
 
 #include "global/types.h"
 
+#include <libtrx/game/gun/rifle.h>
+
 void Gun_Rifle_Draw(LARA_GUN_TYPE weapon_type);
 void Gun_Rifle_Undraw(LARA_GUN_TYPE weapon_type);
-void Gun_Rifle_DrawMeshes(LARA_GUN_TYPE weapon_type);
-void Gun_Rifle_UndrawMeshes(LARA_GUN_TYPE weapon_type);
 void Gun_Rifle_Ready(LARA_GUN_TYPE weapon_type);
 void Gun_Rifle_Control(LARA_GUN_TYPE weapon_type);
 void Gun_Rifle_Animate(LARA_GUN_TYPE weapon_type);
