+++ src/tr2/game/gun/gun_rifle.h
@@ -2,8 +2,8 @@
 
 #include "global/types.h"
 
-void Gun_Rifle_DrawMeshes(LARA_GUN_TYPE weapon_type);
-void Gun_Rifle_UndrawMeshes(LARA_GUN_TYPE weapon_type);
+#include <libtrx/game/gun/rifle.h>
+
 void Gun_Rifle_Ready(LARA_GUN_TYPE weapon_type);
 void Gun_Rifle_Control(LARA_GUN_TYPE weapon_type);
 void Gun_Rifle_FireShotgun(void);
